// backend/src/models/Data.ts
export class Data{
    id: number = 0;
    nome: string = '';
    data_lancamento: string = '';
    descricao: string = '';
    idioma: string = '';
    nota: number = 0;
    genero: number[] = [];
    imagem: string = 'https://image.tmdb.org/t/p/original';
    background_image: string = 'https://image.tmdb.org/t/p/original';
}